extend size of iterations #3

+++ util/runner.zig
@@ -21,7 +21,7 @@ const State = union(enum) {
 
         /// Maximum number of iterations the benchmark can run. This limit helps
         /// to avoid excessively long benchmark runs.
-        max_iterations: u16,
+        max_iterations: u32,
 
         /// Time budget for the benchmark in nanoseconds. This value is used to
         /// determine how long a single benchmark should be allowed to run
